Cache view model changes in dashboard
Resolves #323
Resolves #324

The idea here is,
We start task to run Watch over kubernetes service which compute view model caches and update the current snapshot and writes it to a channel.
Another task reads from channel, adds item to a list (which remembers all the change notifications seen so far) and writes it to additional channels which could be listening.
Get works by returning the current snapshot.
Watch takes snapshot of current change list and adds channel in listening channels. Enumerator first returns items from snapshot and once snapshot is finished, return items from channel as it comes. Clean-up closes the additional channel and removes it from listening channels.

The internal watcher also starts a docker inspect task when we first time see container which has container id and a status. Eventually sending a change notification to update env vars.
Future creation of any view model from kubernetes watch will use the already retrieved env vars.
