+++ src/Aspire.Dashboard/Components/Pages/Containers.razor
@@ -71,7 +71,7 @@
 </div>
 
 @code {
-    protected override Task<List<ContainerViewModel>> GetResources(IDashboardViewModelService dashboardViewModelService)
+    protected override ValueTask<List<ContainerViewModel>> GetResources(IDashboardViewModelService dashboardViewModelService)
         => dashboardViewModelService.GetContainersAsync();
 
     protected override IAsyncEnumerable<ResourceChanged<ContainerViewModel>> WatchResources(
