+++ src/Aspire.Dashboard/Components/Pages/Executables.razor
@@ -81,7 +81,7 @@
 </div>
 
 @code {
-    protected override Task<List<ExecutableViewModel>> GetResources(IDashboardViewModelService dashboardViewModelService)
+    protected override ValueTask<List<ExecutableViewModel>> GetResources(IDashboardViewModelService dashboardViewModelService)
         => dashboardViewModelService.GetExecutablesAsync();
 
     protected override IAsyncEnumerable<ResourceChanged<ExecutableViewModel>> WatchResources(
