+++ src/Aspire.Dashboard/Components/Pages/ProjectLogs.razor
@@ -18,7 +18,7 @@
     protected override string UrlPrefix => "/ProjectLogs";
     protected override string SelectResourceTitle => "Select a Project";
 
-    protected override Task<List<ProjectViewModel>> GetResources(IDashboardViewModelService dashboardViewModelService)
+    protected override ValueTask<List<ProjectViewModel>> GetResources(IDashboardViewModelService dashboardViewModelService)
         => dashboardViewModelService.GetProjectsAsync();
 
     protected override IAsyncEnumerable<ResourceChanged<ProjectViewModel>> WatchResources(IDashboardViewModelService dashboardViewModelService, IEnumerable<NamespacedName>? initialList, CancellationToken cancellationToken)
