+++ src/Aspire.Dashboard/Model/IDashboardViewModelService.cs
@@ -6,9 +6,9 @@ namespace Aspire.Dashboard.Model;
 public interface IDashboardViewModelService
 {
     public string ApplicationName { get; }
-    public Task<List<ContainerViewModel>> GetContainersAsync();
-    public Task<List<ExecutableViewModel>> GetExecutablesAsync();
-    public Task<List<ProjectViewModel>> GetProjectsAsync();
+    public ValueTask<List<ContainerViewModel>> GetContainersAsync();
+    public ValueTask<List<ExecutableViewModel>> GetExecutablesAsync();
+    public ValueTask<List<ProjectViewModel>> GetProjectsAsync();
     public IAsyncEnumerable<ResourceChanged<ContainerViewModel>> WatchContainersAsync(IEnumerable<NamespacedName>? existingContainers = null, CancellationToken cancellationToken = default);
     public IAsyncEnumerable<ResourceChanged<ExecutableViewModel>> WatchExecutablesAsync(IEnumerable<NamespacedName>? existingExecutables = null, CancellationToken cancellationToken = default);
     public IAsyncEnumerable<ResourceChanged<ProjectViewModel>> WatchProjectsAsync(IEnumerable<NamespacedName>? existingProjects = null, CancellationToken cancellationToken = default);
